user page updates
